Add support for java.lang.Class.{isInstance, isAssignableFrom, isInterface} (#668)
* Add support for java.lang.Class.{isInstance, isAssignableFrom, isInterface}

* Add comment to document magical rtti methods

* Improve tests for isInstance, isAssignableFrom

* Improve java.lang.Class.toString

* Always inline rtti helpers

* Add another comment explaining meaning of is helper

* Add test case for java.lang.Class.isInterface

+++ nir/src/main/scala/scala/scalanative/nir/Rt.scala
@@ -6,5 +6,5 @@ import Type._
 object Rt {
   val Object = Class(Global.Top("java.lang.Object"))
   val String = Class(Global.Top("java.lang.String"))
-  val Type   = Struct(Global.None, Seq(Int, Ptr, Long))
+  val Type   = Struct(Global.None, Seq(Int, Ptr, Byte))
 }
